change type

+++ CannonGame/CannonGame/CannonBallData.swift
@@ -6,15 +6,15 @@
 //  Copyright © 2019 kjy95. All rights reserved.
 //
 
-import Foundation
+import UIKit
 /**
  대포알 정보
  */
 class CannonBallData {
-    let maximumXPoint : Float?
-    let maximumYPoint : Float?
-    var cannonRadian : Float?
-    init(maximumXPoint: Float, maximumYPoint: Float, cannonRadian:Float) {
+    let maximumXPoint : CGFloat?
+    let maximumYPoint : CGFloat?
+    var cannonRadian : CGFloat?
+    init(maximumXPoint: CGFloat, maximumYPoint: CGFloat, cannonRadian: CGFloat) {
         self.maximumXPoint = maximumXPoint
         self.maximumYPoint = maximumYPoint
         self.cannonRadian = cannonRadian
